dasda

+++ esphome/components/sht2x/sht2x.cpp
@@ -13,7 +13,7 @@ static const uint8_t SHT2X_COMMAND_SOFT_RESET = 0xFE;
 
 uint8_t SHT2XComponent::get_firmware_version() {
   uint8_t version = 0;
-  if(!this-write_command(0x84, 0xB8)) {
+  if(!this-write_command((uint8_t)0x84, (uint8_t)0xB8)) {
     this->status_set_warning();
     return 0;
   }
